Fix names, delete partial subs GenericPrincipal and GenericIdentity
